# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Auspac Documentation",
        "description": "Project Inventory Auspac Laravel 12 Documentation",
        "contact": {
            "email": "example@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/taxcore/status": {
            "get": {
                "tags": [
                    "TaxCore"
                ],
                "summary": "Check TaxCore connection status",
                "description": "Reads the installed PFX certificate, extracts the API endpoint, and verifies that a token can be obtained from the V-SDC.",
                "operationId": "905d42ee9422206bae6320ca6685a92b",
                "responses": {
                    "200": {
                        "description": "Connection is healthy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "TaxCore connection OK"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "properties": {
                                                "endpoint": {
                                                    "type": "string"
                                                },
                                                "token_preview": {
                                                    "type": "string"
                                                },
                                                "certificate": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Connection failed"
                    }
                }
            }
        },
        "/api/taxcore/sync-rates": {
            "post": {
                "tags": [
                    "TaxCore"
                ],
                "summary": "Sync tax rates from TaxCore",
                "description": "Calls Get Environment Parameters on the V-SDC and persists the returned tax rates in the local database.",
                "operationId": "35d5193a26306eafcf507d1764e4bc27",
                "responses": {
                    "200": {
                        "description": "Tax rates synced successfully"
                    },
                    "503": {
                        "description": "Sync failed"
                    }
                }
            }
        },
...
        "/api/taxcore/fiscalize": {
            "post": {
                "tags": [
                    "TaxCore"
                ],
                "summary": "Fiscalize (send) an invoice to TaxCore",
                "description": "Builds the TaxCore invoice payload from the request body, sends it to the V-SDC, and persists the fiscal response (QR code, digital signature, verification URL).",
                "operationId": "32a9e97345815a89883831bf299e5148",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoiceNumber",
                                    "invoiceDate",
                                    "items"
                                ],
                                "properties": {
                                    "invoiceNumber": {
                                        "type": "string",
                                        "example": "INV-0001"
                                    },
                                    "invoiceDate": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "saleId": {
                                        "type": "string",
                                        "example": "42"
                                    },
                                    "buyerName": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "buyerTaxId": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Product A"
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": 2
                                                },
                                                "unitPrice": {
                                                    "type": "number",
                                                    "example": 10
                                                },
                                                "taxCode": {
                                                    "type": "string",
                                                    "example": "A"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Invoice queued for fiscalization"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/taxcore/invoices/{id}/retry": {
            "post": {
                "tags": [
                    "TaxCore"
                ],
                "summary": "Retry a dead-lettered or failed invoice",
                "operationId": "3178f4bd0120c570d7ad6d69872ee6dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Invoice re-queued"
                    },
                    "404": {
                        "description": "Invoice not found"
                    },
                    "409": {
                        "description": "Invoice is not in a retryable state"
                    }
                }
            }
        },
        "/api/taxcore/invoices": {
            "get": {
                "tags": [
                    "TaxCore"
                ],
                "summary": "List all fiscalized invoices",
                "description": "Returns a paginated list of all invoices that have been fiscalized and stored locally.",
                "operationId": "b6456a3f044c92b560af6e7ff45e5e70",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of fiscal invoices"
                    }
                }
            }
        },
        "/api/taxcore/invoices/{id}": {
            "get": {
                "tags": [
                    "TaxCore"
                ],
                "summary": "Get a single fiscalized invoice",
                "operationId": "2fd93a1e2e9a3a0a59a97ddaca80d392",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice found"
                    },
                    "404": {
                        "description": "Invoice not found"
                    }
                }
            }
        },
        "/api/taxcore/test-invoice": {
            "post": {
                "tags": [
                    "TaxCore"
                ],
                "summary": "Send a hardcoded test invoice to TaxCore",
                "description": "Sends a sample Normal Sale invoice with one product and returns the full raw fiscal response including QR code, signature and verification URL.",
                "operationId": "ce24ce3a556922011f3acf756e5fcba7",
                "responses": {
                    "200": {
                        "description": "Test invoice fiscalized successfully"
                    },
                    "503": {
                        "description": "Fiscalization failed"
                    }
                }
            }
        },
...
    },
...
}